requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Task Queue
celery==5.3.6
redis==5.0.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# AI/ML
openai==1.10.0
anthropic==0.9.0
tiktoken==0.5.2

# Voice Synthesis
elevenlabs==0.2.27
google-cloud-texttospeech==2.16.0

# Video Processing
moviepy==1.0.3
ffmpeg-python==0.2.0
Pillow==10.2.0
numpy==1.26.3

# Image Generation/Processing
stability-sdk==0.8.4
opencv-python==4.9.0.80

# News & Content Collection
feedparser==6.0.11
newspaper3k==0.2.8
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0

# YouTube API
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Trend Analysis
pytrends==4.9.2

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.6
httpx==0.26.0
aiohttp==3.9.1
tenacity==8.2.3

# Logging & Monitoring
loguru==0.7.2
sentry-sdk==1.40.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0

# Scheduler
APScheduler==3.10.4

# Configuration
pyyaml==6.0.1
